Huge refactor to remove unused stuff
