--- src/board/mod.rs
@@ -1,5 +1,0 @@
-mod board;
-pub mod movements;
-
-pub use board::*;
-pub use movements::{Direction, ADJACENCY_TABLE};
